llm_crud/requirements.txt
mysql-connector-python==9.0.0
openai==1.60.1
python-dotenv==1.0.1
pyyaml==6.0.2
slack_bolt==1.22.0